app/requirements.txt
fastapi~=0.85.2
pydantic~=1.10.2
uvicorn~=0.18.3
tortoise-orm~=0.19.2
docker~=6.0.1
fastapi-utils~=0.2.1
httpx~=0.23.0
PyJWT~=2.6.0
psutil~=5.9.3
aiofiles~=22.1.0
aiomysql~=0.1.1
iso8601~=1.1.0